Set up join workspace page
