+++ frontend/src/pages/Auth/JoinWorkspace.tsx
@@ -1,194 +1,143 @@
-import { ApolloError } from '@apollo/client'
-import { Button } from '@components/Button'
+import { Form, Stack, Text, useFormState } from '@highlight-run/ui'
+import { message } from 'antd'
+import { useEffect } from 'react'
+import { useNavigate } from 'react-router-dom'
+import { useLocalStorage } from 'react-use'
+
+import { getEmailDomain } from '@/components/AutoJoinEmailsInput'
+import { Button } from '@/components/Button'
 import {
 	AppLoadingState,
 	useAppLoadingContext,
-} from '@context/AppLoadingContext'
+} from '@/context/AppLoadingContext'
 import {
-	useAddAdminToWorkspaceMutation,
-	useGetWorkspaceForInviteLinkQuery,
+	useGetAdminQuery,
 	useGetWorkspacesQuery,
-} from '@graph/hooks'
-import { Box, Callout, Stack, Text } from '@highlight-run/ui'
-import { AuthBody, AuthFooter, AuthHeader } from '@pages/Auth/Layout'
-import { Landing } from '@pages/Landing/Landing'
-import useLocalStorage from '@rehooks/local-storage'
-import { message } from 'antd'
-import { H } from 'highlight.run'
-import React, { useEffect } from 'react'
-import { Navigate, useMatch, useNavigate } from 'react-router-dom'
+	useJoinWorkspaceMutation,
+} from '@/graph/generated/hooks'
+import { AuthBody, AuthFooter, AuthHeader } from '@/pages/Auth/Layout'
+import { Landing } from '@/pages/Landing/Landing'
 
-import * as styles from './AuthRouter.css'
+import * as styles from './AdminForm.css'
+import * as authRouterStyles from './AuthRouter.css'
+
+export const DISMISS_JOIN_WORKSPACE_LOCAL_STORAGE_KEY =
+	'highlightDismissedJoinWorkspace'
 
 export const JoinWorkspace = () => {
-	const [_, setInviteCode] = useLocalStorage('highlightInviteCode')
-	const joinWorkspaceMatch = useMatch('/invite/:inviteCode')
-	const inviteCode = joinWorkspaceMatch?.params.inviteCode
-	const { setLoadingState } = useAppLoadingContext()
+	const { data, loading } = useGetWorkspacesQuery()
+	const { data: adminData } = useGetAdminQuery()
 	const navigate = useNavigate()
-	const { data, error, loading } = useGetWorkspaceForInviteLinkQuery({
-		variables: {
-			secret: inviteCode!,
-		},
-	})
-	const [
-		addAdminToWorkspace,
-		{ loading: addAdminLoading, error: addAdminError },
-	] = useAddAdminToWorkspaceMutation()
-	const { data: workspacesData, loading: workspacesLoading } =
-		useGetWorkspacesQuery()
-	const workspaceId = data?.workspace_for_invite_link?.workspace_id
-	const workspaceName = data?.workspace_for_invite_link?.workspace_name
-	const alreadyInWorkspace = workspacesData?.workspaces?.some(
-		(w) => w?.id && w.id === workspaceId,
+	const { setLoadingState } = useAppLoadingContext()
+	const [joinWorkspace, { loading: joinLoading }] = useJoinWorkspaceMutation()
+	const [_, setDismissedJoinWorkspace] = useLocalStorage(
+		DISMISS_JOIN_WORKSPACE_LOCAL_STORAGE_KEY,
+		false,
 	)
 
-	const clearInviteAndRedirect = () => {
-		setInviteCode('')
-		navigate('/')
-	}
+	const form = useFormState({
+		defaultValues: {
+			workspaceId: '',
+		},
+	})
+
+	form.useSubmit(async () => {
+		const response = await joinWorkspace({
+			variables: {
+				workspace_id: form.values.workspaceId,
+			},
+		})
+
+		if (!!response.data?.joinWorkspace) {
+			message.success('Successfuly joined workspace!', 1)
+			navigate('/about_you', { replace: true })
+		}
+	})
 
 	useEffect(() => {
-		setLoadingState(AppLoadingState.LOADED)
-	}, [setLoadingState])
-
-	useEffect(() => {
-		if (error) {
-			clearInviteAndRedirect()
-
-			if (error.message.indexOf('expired') > -1) {
-				message.error(
-					'This invite link has expired. Please ask your admin for a new one.',
-				)
+		if (!loading) {
+			if (!data?.joinable_workspaces?.length) {
+				navigate('/about_you', { replace: true })
 			} else {
-				message.error('Invalid invite code.')
+				setLoadingState(AppLoadingState.LOADED)
 			}
 		}
+	}, [data?.joinable_workspaces?.length, loading, navigate, setLoadingState])
 
-		if (alreadyInWorkspace) {
-			clearInviteAndRedirect()
-			message.success('You are already a member of this workspace.')
-		}
-		// eslint-disable-next-line react-hooks/exhaustive-deps
-	}, [alreadyInWorkspace, error])
-
-	if (loading || workspacesLoading) {
+	if (loading) {
 		return null
 	}
 
-	if (!workspaceId) {
-		return <Navigate replace to="/" />
-	}
+	const emailDomain = getEmailDomain(adminData?.admin?.email)
 
 	return (
 		<Landing>
-			<Box cssClass={styles.container}>
+			<Form className={authRouterStyles.container} state={form}>
 				<AuthHeader>
-					<Text color="moderate">Accept Invitation?</Text>
+					<Text color="moderate">Join Workspace</Text>
 				</AuthHeader>
 				<AuthBody>
-					<Stack gap="20">
-						<Text align="center">
-							Do you want to accept the invitation to join "
-							{workspaceName}"?
+					<Stack gap="16" direction="column">
+						<Text>
+							Based on your <b>@{emailDomain}</b> email address
+							you are able to join the following workspaces.
 						</Text>
-						{!!addAdminError && (
-							<Callout
-								kind="error"
-								title={getAlertMessage(addAdminError).title}
-							>
-								{getAlertMessage(addAdminError).description}
-							</Callout>
-						)}
+
+						<select
+							className={styles.select}
+							value={form.values.workspaceId}
+							onChange={(e) => {
+								const selectedWorkspace =
+									data?.joinable_workspaces?.find(
+										(workspace) =>
+											workspace?.id === e.target.value,
+									)
+
+								form.setValue(
+									form.names.workspaceId,
+									selectedWorkspace?.id,
+								)
+							}}
+						>
+							<option value="" disabled>
+								Select a workspace
+							</option>
+
+							{/* TODO: Clean up the select UI */}
+							{data?.joinable_workspaces?.map((workspace) => (
+								<option
+									key={workspace?.id}
+									value={workspace?.id}
+								>
+									{workspace?.name}
+								</option>
+							))}
+						</select>
 					</Stack>
 				</AuthBody>
 				<AuthFooter>
-					<Stack gap="8">
-						<Button
-							trackingId="join-workspace-accept"
-							kind="primary"
-							loading={addAdminLoading}
-							onClick={async () => {
-								try {
-									await addAdminToWorkspace({
-										variables: {
-											workspace_id: workspaceId,
-											invite_id: inviteCode!,
-										},
-									})
-
-									message.success(
-										`Successfully joined workspace "${workspaceName}"!`,
-									)
-
-									clearInviteAndRedirect()
-								} catch (_e) {
-									message.error(
-										'Failed to join the workspace. Please try again.',
-									)
-									window.Intercom(
-										'showNewMessage',
-										`I'm having trouble joining the "${workspaceName}" workspace....`,
-									)
-								}
-							}}
-						>
-							Join
-						</Button>
-						<Button
-							kind="secondary"
-							trackingId="join-workspace-ignore"
-							onClick={() => {
-								setInviteCode('ignored')
-								navigate('/')
-							}}
-						>
-							Ignore invitation
-						</Button>
-					</Stack>
+					<Button
+						type="submit"
+						kind="primary"
+						disabled={!form.values.workspaceId}
+						onClick={() => null}
+						trackingId="join-workspace_submit"
+						loading={joinLoading}
+					>
+						Join Workspace
+					</Button>
+					<Button
+						kind="secondary"
+						trackingId="join-workspace_skip"
+						onClick={() => {
+							setDismissedJoinWorkspace(true)
+							navigate('/about_you', { replace: true })
+						}}
+					>
+						<Text color="moderate">Cancel</Text>
+					</Button>
 				</AuthFooter>
-			</Box>
+			</Form>
 		</Landing>
 	)
-}
-
-const getAlertMessage = (
-	error: ApolloError,
-): { title: string; description: string } => {
-	const { message } = error
-	const defaultAlertProps = {
-		title: 'A problem occurred while trying to join the project.',
-		description:
-			'This is usually an intermittent issue. If this keeps happening please reach out to us. We are probably already looking into it!',
-	}
-	const proxyError = new Error(
-		'A 500 occurred when an admin tried joining a workspace.',
-	)
-
-	if (message.includes('403')) {
-		return {
-			title: defaultAlertProps.title,
-			description:
-				"It doesn't look like this is a valid invite link. Ask the person that shared this link with you to double check.",
-		}
-	}
-
-	if (message.includes('404')) {
-		return {
-			title: "This invite doesn't exist",
-			description:
-				'Ask the person that shared this link with you to re-invite you. An invite link expires after it is used.',
-		}
-	}
-
-	if (message.includes('405')) {
-		return {
-			title: 'The invite link has expired',
-			description:
-				'Ask the person that shared with you the invite to create a new invite for you.',
-		}
-	}
-
-	H.consumeError(proxyError)
-	return defaultAlertProps
 }
